[WIP] rename Python file from color_manager.py to color_manager_curses.py
