+++ suplemon/color_manager_curses.py
